requirements.txt
# --- Core Server ---
flask
flask-cors
python-dotenv
uvicorn
fastapi

# --- AI & Data Processing ---
groq
httpx
pydantic
pandas
openpyxl
markdown

# --- Automation & Scraping ---
selenium
undetected-chromedriver
webdriver-manager
requests
beautifulsoup4
lxml

# --- Security & Auth ---
pyjwt

# --- Windows Hardware Control (Printer) ---
# Note: This only works on Windows
pywin32